Moved random line generator to util package and cleaned it up a bit
